Version 0.1.0
- Add option to enable clipper library for benchmarks: `CAVC_ENABLE_CLIPPER_BENCHMARKS`. If enabled, clipper will be fetched and built automatically. default is disabled.
- Remove unused includes
- Rename of some variables declaration shadow the local/previous variable
  - if parameter shadows the local variable, the local variable is renamed to `p_<old_name>`
  - if variable in function shadows the local variable, the parameter is renamed to `l_<old_name>`
- Format code in tests

+++ tests/benchmarks/combinebenchmarks.cpp
@@ -21,7 +21,8 @@ struct CombineShiftedSetup {
     std::size_t shiftedCount = 16;
     shiftedProfiles.reserve(shiftedCount);
     for (std::size_t i = 0; i < shiftedCount; ++i) {
-      double angle = static_cast<double>(i) / shiftedCount * cavc::utils::tau<double>();
+      double angle =
+          static_cast<double>(i) / static_cast<double>(shiftedCount) * cavc::utils::tau<double>();
       shiftedProfiles.push_back(createShifted(angle));
     }
   }
